Revisions to LabVIEW velocity closed loop.  -- NOT TESTED
I attempted to make revisions to the LabVIEW closed loop example to use the VI that are in the palette.

Also moved visual and user interaction to Robot Main to allow this to run easier on the robot interactively.

I've not used the Spark Max so I don't have experience with this or any way to test.